Massively optimize shot collisions
When calculating shot collisions for player fired attacks, only look for
enemies at most a screen away from the player. Also, remove special
treatment for colliding shots with debris and closed shells. As a side
effect, shooting closed shells will push them in the direction of the
shot. It's not a bug, it's a feature!